multiline function declaration arguments when needed
Implement tryMaxColumns which first attempt to print whatever is in callback a() making sure that maxColumns is Infinity to make it only one line. If this fails, then it restores the state of the printer to whatever it was before this attempt and then prints option b().

I'm not 100% confident that my checkpoint/restore covers all the cases as there is a lot of mutable state in those files, but it works well in my limited tests.